+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 
-## 0.1.1-SNAPSHOT
+## 0.1.2-SNAPSHOT
+
+- ?
+
+## 0.1.1
 
 - #1 Load printing namespaces for EDN
 - #2 add more types: Agent Volatile Namespace Path InetAddress